+++ buildVars.py
@@ -31,7 +31,8 @@ NVDA menu>preferences>inputGestures>Link Library."""),
 	addon_version="2.2.8",
 	# Brief changelog for this version
 	# Translators: what's new content for the add-on version to be shown in the add-on store
-	addon_changelog=_("""
+	addon_changelog=_("""- Fix a bug in 'Show source url' button in link dialog, that appeared in languages other than English.
+- Add Spanish documentation, contributed by [Ana Giselle](https://github.com/ana-giselle-ar86).
 """),
 	# Author(s)
 	addon_author="ibrahim hamadeh<ibra.hamadeh@hotmail.com>",